+++ tests/test-suites/Atoms/Checkbox.testsuite.tsx
@@ -7,7 +7,7 @@ const onChangeStub = Sinon.stub();
 
 test("rendered without errors - no params", () => {
   const wrapper = mount(<Checkbox />);
-  wrapper.find("#modular-checkbox.check").simulate("click");
+  wrapper.find("#modular-checkbox").simulate("click");
   expect(onChangeStub).not.toBeCalled;
 });
 
@@ -15,6 +15,6 @@ test("rendered without errors", () => {
   const wrapper = mount(
     <Checkbox value={true} onChange={onChangeStub} onClick={onChangeStub} />
   );
-  wrapper.find("#modular-checkbox.check").simulate("click");
+  wrapper.find("#modular-checkbox").simulate("click");
   expect(onChangeStub).toBeCalled;
 });